Tolerate RSASSA-PSS in OpenJCEPlus signature
The mask generation function (MGF) based on a hash algorithm
is recommended to use the same hash function as the hash function
fingerprinting the message for RSASSA-PSS. However, the structures
in [PKCS#1v2.1] allow for separate parameterization of the MGF
and the message digest.

OpenJCEPlus restricts to use the same hash algorithm for both MGF
and message digest. This will cause test failures since the tests
are expecting different MGF and message digest behavior.

Signed-off-by: Jinhang Zhang <Jinhang.Zhang@ibm.com>