Add key algorithm check for native CBC
Native implementation of CBC mode is used as default while
initializing an AES cipher instance. A secret key will be used
to initilize an AES cipher instance. However, there is a lack
of check to see if the implementation algorithm corresponding
to the secret key is consistent with the cipher's algorithm.

This PR solves the issue #21886